Add location

+++ ThePersonalTrainerClub/ThePersonalTrainerClub/Model/Model/LocationModel.swift
@@ -11,6 +11,6 @@ import Foundation
 struct LocationModel: Codable {
     let id: String
     let type: String
-    let coordinates: [Float]
+    let coordinates: [Double]
     let description: String
 }
